Delete backups older than N days.
This changes the API of the BackupSheet class for deleting old backups:
We now specify a number of days instead of a specific date until which
to truncate old backups. This allows us to keep the code as simple as
possible in main.js, by removing the need to manipulate Date objects.