Refactor: Add sheet.appendRows().
This function will come in handy to implement daily backups where the
data is simply overwritten without including a backup date field.
This change also eliminates the bad practice of appending rows one by
one in the overwrite() function, replacing it with the better practice
of bulk-writing all the data at once.