Simplify getColumnId() and getHeader() implementations.
Modify getHeader() to return an Array<string> instead of a Range object,
because I can't think of a case where we'd need to do anything else than
.getValues()[0] on the returned range.

The new implementation of getColumnId() relies only on the Javascript
builtin Array.indexOf. This change also affects the return value for
the case where the column doesn't exist.