Make modules compatible with Apps Script
requirejs is used by tests (implemented using Jest).

However, requirejs isn't present in the Apps Script environment.
Unlike in node, all .gs files are executed thus functions, classes
and variables are global. This change will allow script.js to be
executed as-is, without choking on the `export` or `module`
keywords in Apps Script.