Add 2 computed columns to the TTTR matching metrics sheet.
Tracking the time that a project has spent in verification isn't
sufficient: we need to also track the number of days it may have spent
waiting in "Project form submitted" stage.

This change collects 3 metrics instead of 1. We may not need all three
in the dashboard, but we haven't yet decided which charts to display, so
having the historical data may come in handy once we settle on a design.

The "Days waiting or in verification" computation could also be dnone in
Data Studio instead of here, but having it here allows this code to be
versioned.